refactor: extract core state from AgentWorkspace
